refactor(levm): refactor Opcodes LEVM (#712)
**Motivation**

<!-- Why does this pull request exist? What are its goals? -->
Modularize code, the file `vm.rs` has turned huge after the
implementation of many of the opcodes, the idea is for each opcode group
to have their own file.
In addition to this, the error handling has been modified. This emerged
as a necessity for keeping the code more clear and concise. Tests were
modified accordingly to suit the changes made.

**Description**

<!-- A clear and concise general description of the changes this PR
introduces -->
- Split opcodes into different files, grouping them by categories
defined in [Ethereum Yellow
Paper](https://ethereum.github.io/yellowpaper/paper.pdf)
- Now `vm.execute()` returns an `ExecutionResult` instead of
`Result<ExecutionResult, VMError>`
- `VMError::FatalError` is no longed considered as a possible error.

**Considerations**
- When moving an opcode that has a `continue` into a function I replace
that for a `return` so that the behavior stays the same.
- Some methods requiere access to `self`, and some other not. For
simplicity I send `&self` in every single method call. Example:
`self.op_add(...)` (I'm implicity passing `&self` as parameter)
- `current_call_frame` is sent as parameter to opcodes methods even
though it can be accessed within `self` because it is very common in
many opcodes.
- Added a new type `OpcodeSuccess` for better handling of success cases
in the execution of opcodes.
- A few tests were removed because they were not necessary anymore
- `Revert` case never happens because when an error happens the Halting
is exceptional, so it consumes all gas. The Reverts from natural halting
shouldn't have a VMError associated to them.

<!-- Link to issues: Resolves #111, Resolves #222 -->

**Concerns**
There are various opcodes (specially bit operations and arithmetic ones)
that lack tests. They were added in PR #696 though.

Closes #684
