+++ crates/vm/levm/src/lib.rs
@@ -2,6 +2,7 @@ pub mod block;
 pub mod call_frame;
 pub mod constants;
 pub mod memory;
+pub mod opcode_handlers;
 pub mod opcodes;
 pub mod operations;
 pub mod primitives;
