+++ crates/vm/levm/src/opcode_handlers/stop_and_arithmetic.rs
@@ -1,0 +1,298 @@
+use crate::vm_result::ResultReason;
+
+// Stop and Arithmetic Operations (12)
+// Opcodes: STOP, ADD, SUB, MUL, DIV, SDIV, MOD, SMOD, ADDMOD, MULMOD, EXP, SIGNEXTEND
+use super::*;
+
+impl VM {
+    // STOP operation
+    pub fn op_stop(
+        &mut self,
+        current_call_frame: &mut CallFrame,
+    ) -> Result<OpcodeSuccess, VMError> {
+        self.call_frames.push(current_call_frame.clone());
+        Ok(OpcodeSuccess::Result(ResultReason::Stop))
+    }
+    // ADD operation
+    pub fn op_add(&mut self, current_call_frame: &mut CallFrame) -> Result<OpcodeSuccess, VMError> {
+        if self.env.consumed_gas + gas_cost::ADD > self.env.gas_limit {
+            return Err(VMError::OutOfGas);
+        }
+        let augend = current_call_frame.stack.pop()?;
+        let addend = current_call_frame.stack.pop()?;
+        let sum = augend.overflowing_add(addend).0;
+        current_call_frame.stack.push(sum)?;
+        self.env.consumed_gas += gas_cost::ADD;
+
+        Ok(OpcodeSuccess::Continue)
+    }
+
+    // SUB operation
+    pub fn op_sub(&mut self, current_call_frame: &mut CallFrame) -> Result<OpcodeSuccess, VMError> {
+        if self.env.consumed_gas + gas_cost::SUB > self.env.gas_limit {
+            return Err(VMError::OutOfGas);
+        }
+        let minuend = current_call_frame.stack.pop()?;
+        let subtrahend = current_call_frame.stack.pop()?;
+        let difference = minuend.overflowing_sub(subtrahend).0;
+        current_call_frame.stack.push(difference)?;
+        self.env.consumed_gas += gas_cost::SUB;
+
+        Ok(OpcodeSuccess::Continue)
+    }
+
+    // MUL operation
+    pub fn op_mul(&mut self, current_call_frame: &mut CallFrame) -> Result<OpcodeSuccess, VMError> {
+        if self.env.consumed_gas + gas_cost::MUL > self.env.gas_limit {
+            return Err(VMError::OutOfGas);
+        }
+        let multiplicand = current_call_frame.stack.pop()?;
+        let multiplier = current_call_frame.stack.pop()?;
+        let product = multiplicand.overflowing_mul(multiplier).0;
+        current_call_frame.stack.push(product)?;
+        self.env.consumed_gas += gas_cost::MUL;
+
+        Ok(OpcodeSuccess::Continue)
+    }
+
+    // DIV operation
+    pub fn op_div(&mut self, current_call_frame: &mut CallFrame) -> Result<OpcodeSuccess, VMError> {
+        if self.env.consumed_gas + gas_cost::DIV > self.env.gas_limit {
+            return Err(VMError::OutOfGas);
+        }
+        let dividend = current_call_frame.stack.pop()?;
+        let divisor = current_call_frame.stack.pop()?;
+        if divisor.is_zero() {
+            current_call_frame.stack.push(U256::zero())?;
+            return Ok(OpcodeSuccess::Continue);
+        }
+        let quotient = dividend / divisor;
+        current_call_frame.stack.push(quotient)?;
+        self.env.consumed_gas += gas_cost::DIV;
+
+        Ok(OpcodeSuccess::Continue)
+    }
+
+    // SDIV operation
+    pub fn op_sdiv(
+        &mut self,
+        current_call_frame: &mut CallFrame,
+    ) -> Result<OpcodeSuccess, VMError> {
+        if self.env.consumed_gas + gas_cost::SDIV > self.env.gas_limit {
+            return Err(VMError::OutOfGas);
+        }
+        let dividend = current_call_frame.stack.pop()?;
+        let divisor = current_call_frame.stack.pop()?;
+        if divisor.is_zero() {
+            current_call_frame.stack.push(U256::zero())?;
+            return Ok(OpcodeSuccess::Continue);
+        }
+
+        let dividend_is_negative = is_negative(dividend);
+        let divisor_is_negative = is_negative(divisor);
+        let dividend = if dividend_is_negative {
+            negate(dividend)
+        } else {
+            dividend
+        };
+        let divisor = if divisor_is_negative {
+            negate(divisor)
+        } else {
+            divisor
+        };
+        let quotient = dividend / divisor;
+        let quotient_is_negative = dividend_is_negative ^ divisor_is_negative;
+        let quotient = if quotient_is_negative {
+            negate(quotient)
+        } else {
+            quotient
+        };
+
+        current_call_frame.stack.push(quotient)?;
+        self.env.consumed_gas += gas_cost::SDIV;
+
+        Ok(OpcodeSuccess::Continue)
+    }
+
+    // MOD operation
+    pub fn op_mod(&mut self, current_call_frame: &mut CallFrame) -> Result<OpcodeSuccess, VMError> {
+        if self.env.consumed_gas + gas_cost::MOD > self.env.gas_limit {
+            return Err(VMError::OutOfGas);
+        }
+        let dividend = current_call_frame.stack.pop()?;
+        let divisor = current_call_frame.stack.pop()?;
+        if divisor.is_zero() {
+            current_call_frame.stack.push(U256::zero())?;
+            return Ok(OpcodeSuccess::Continue);
+        }
+        let remainder = dividend % divisor;
+        current_call_frame.stack.push(remainder)?;
+        self.env.consumed_gas += gas_cost::MOD;
+
+        Ok(OpcodeSuccess::Continue)
+    }
+
+    // SMOD operation
+    pub fn op_smod(
+        &mut self,
+        current_call_frame: &mut CallFrame,
+    ) -> Result<OpcodeSuccess, VMError> {
+        if self.env.consumed_gas + gas_cost::SMOD > self.env.gas_limit {
+            return Err(VMError::OutOfGas);
+        }
+        let dividend = current_call_frame.stack.pop()?;
+        let divisor = current_call_frame.stack.pop()?;
+        if divisor.is_zero() {
+            current_call_frame.stack.push(U256::zero())?;
+        } else {
+            let dividend_is_negative = is_negative(dividend);
+            let divisor_is_negative = is_negative(divisor);
+
+            let dividend = if dividend_is_negative {
+                negate(dividend)
+            } else {
+                dividend
+            };
+            let divisor = if divisor_is_negative {
+                negate(divisor)
+            } else {
+                divisor
+            };
+
+            let remainder = dividend % divisor;
+            let remainder_is_negative = dividend_is_negative ^ divisor_is_negative;
+
+            let remainder = if remainder_is_negative {
+                negate(remainder)
+            } else {
+                remainder
+            };
+
+            current_call_frame.stack.push(remainder)?;
+        }
+        self.env.consumed_gas += gas_cost::SMOD;
+
+        Ok(OpcodeSuccess::Continue)
+    }
+
+    // ADDMOD operation
+    pub fn op_addmod(
+        &mut self,
+        current_call_frame: &mut CallFrame,
+    ) -> Result<OpcodeSuccess, VMError> {
+        if self.env.consumed_gas + gas_cost::ADDMOD > self.env.gas_limit {
+            return Err(VMError::OutOfGas);
+        }
+        let augend = current_call_frame.stack.pop()?;
+        let addend = current_call_frame.stack.pop()?;
+        let divisor = current_call_frame.stack.pop()?;
+        if divisor.is_zero() {
+            current_call_frame.stack.push(U256::zero())?;
+            return Ok(OpcodeSuccess::Continue);
+        }
+        let (sum, overflow) = augend.overflowing_add(addend);
+        let mut remainder = sum % divisor;
+        if overflow || remainder > divisor {
+            remainder = remainder.overflowing_sub(divisor).0;
+        }
+
+        current_call_frame.stack.push(remainder)?;
+        self.env.consumed_gas += gas_cost::ADDMOD;
+
+        Ok(OpcodeSuccess::Continue)
+    }
+
+    // MULMOD operation
+    pub fn op_mulmod(
+        &mut self,
+        current_call_frame: &mut CallFrame,
+    ) -> Result<OpcodeSuccess, VMError> {
+        if self.env.consumed_gas + gas_cost::MULMOD > self.env.gas_limit {
+            return Err(VMError::OutOfGas);
+        }
+        let multiplicand = U512::from(current_call_frame.stack.pop()?);
+        let multiplier = U512::from(current_call_frame.stack.pop()?);
+        let divisor = U512::from(current_call_frame.stack.pop()?);
+        if divisor.is_zero() {
+            current_call_frame.stack.push(U256::zero())?;
+            return Ok(OpcodeSuccess::Continue);
+        }
+
+        let (product, overflow) = multiplicand.overflowing_mul(multiplier);
+        let mut remainder = product % divisor;
+        if overflow || remainder > divisor {
+            remainder = remainder.overflowing_sub(divisor).0;
+        }
+        let mut result = Vec::new();
+        for byte in remainder.0.iter().take(4) {
+            let bytes = byte.to_le_bytes();
+            result.extend_from_slice(&bytes);
+        }
+        // before reverse we have something like [120, 255, 0, 0....]
+        // after reverse we get the [0, 0, ...., 255, 120] which is the correct order for the little endian u256
+        result.reverse();
+        let remainder = U256::from(result.as_slice());
+        current_call_frame.stack.push(remainder)?;
+        self.env.consumed_gas += gas_cost::MULMOD;
+
+        Ok(OpcodeSuccess::Continue)
+    }
+
+    // EXP operation
+    pub fn op_exp(&mut self, current_call_frame: &mut CallFrame) -> Result<OpcodeSuccess, VMError> {
+        let base = current_call_frame.stack.pop()?;
+        let exponent = current_call_frame.stack.pop()?;
+
+        let exponent_byte_size = (exponent.bits() as u64 + 7) / 8;
+        let gas_cost = gas_cost::EXP_STATIC + gas_cost::EXP_DYNAMIC_BASE * exponent_byte_size;
+        if self.env.consumed_gas + gas_cost > self.env.gas_limit {
+            return Err(VMError::OutOfGas);
+        }
+
+        let power = base.overflowing_pow(exponent).0;
+        current_call_frame.stack.push(power)?;
+        self.env.consumed_gas += gas_cost;
+
+        Ok(OpcodeSuccess::Continue)
+    }
+
+    // SIGNEXTEND operation
+    pub fn op_signextend(
+        &mut self,
+        current_call_frame: &mut CallFrame,
+    ) -> Result<OpcodeSuccess, VMError> {
+        if self.env.consumed_gas + gas_cost::SIGNEXTEND > self.env.gas_limit {
+            return Err(VMError::OutOfGas);
+        }
+        let byte_size = current_call_frame.stack.pop()?;
+        let value_to_extend = current_call_frame.stack.pop()?;
+
+        let bits_per_byte = U256::from(8);
+        let sign_bit_position_on_byte = 7;
+        let max_byte_size = 31;
+
+        let byte_size = byte_size.min(U256::from(max_byte_size));
+        let sign_bit_index = bits_per_byte * byte_size + sign_bit_position_on_byte;
+        let is_negative = value_to_extend.bit(sign_bit_index.as_usize());
+        let sign_bit_mask = (U256::one() << sign_bit_index) - U256::one();
+        let result = if is_negative {
+            value_to_extend | !sign_bit_mask
+        } else {
+            value_to_extend & sign_bit_mask
+        };
+        current_call_frame.stack.push(result)?;
+        self.env.consumed_gas += gas_cost::SIGNEXTEND;
+
+        Ok(OpcodeSuccess::Continue)
+    }
+}
+
+/// Shifts the value to the right by 255 bits and checks the most significant bit is a 1
+fn is_negative(value: U256) -> bool {
+    value.bit(255)
+}
+
+/// Negates a number in two's complement
+fn negate(value: U256) -> U256 {
+    !value + U256::one()
+}
